Package restructuring, remove unnecessary lines
