+++ java/src/main/scala/com/databricks/mlflow/sagemaker/MLeapPredictor.scala
@@ -1,10 +1,7 @@
 package com.databricks.mlflow.sagemaker;
 
-import java.nio.charset.Charset
-
 import ml.combust.bundle.BundleFile
 import ml.combust.mleap.runtime.MleapSupport._
-import ml.combust.mleap.runtime.serialization.FrameReader
 import ml.combust.mleap.runtime.frame.Transformer
 
 import resource._
@@ -16,8 +13,6 @@ class MLeapPredictor(var modelPath : String) extends Predictor {
   }).opt.get
 
   val pipeline = bundle.root
-  val frameReader = FrameReader()
-  val jsonCharset = Charset.forName("UTF-8")
 
   def getPipeline() : Transformer = {
     this.pipeline
